Fix compilation errors in UniLinks implementation
- Fixed ambiguous operator comparison by casting ulong to int
- Added explicit type parameters to extension method calls

+++ Platform/Platform.Data.Core/UniLinksWrapper.cs
@@ -60,7 +60,7 @@ namespace Platform.Data.Core
             // Read: Out operation to extract a specific part
             _links.Out(linkParts =>
             {
-                if (linkParts.Length > partType)
+                if (linkParts.Length > (int)partType)
                 {
                     result = linkParts[partType];
                 }
